+++ character.h
@@ -8,8 +8,8 @@
 class MyCharacter {
 public:
     std::string name;
-    int power, agility, health, defense, luck, gold, xp, level, statsAdded;
-    MyCharacter(std::string& name, int power, int agility, int health, int defense, int luck, int gold, int xp, int level, int statsAdded);
+    double power, agility, health, defense, luck, gold, xp, level, statsAdded;
+    MyCharacter(std::string& name, double power, double agility, double health, double defense, double luck, double gold, double xp, double level, double statsAdded);
     void showStats ();
     void AddStats();
 };
